requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
Faker==4.14.2
python-dateutil==2.8.1
pytz==2020.4
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
validate-docbr==1.8.2
